Update patch set 1
Abandoned

Actually, there is no stable/mitaka branch, so this is not needed. It turned out to be caused by the stable/mitaka worker not building often enough, and it tried to build an older commit with the new spec. 

I'll remove the rpm-mitaka branch, it's not needed

Patch-set: 1
Status: abandoned