Update patch set 1
Patch Set 1:

This is a weird thing. py27 and docs build gate jobs for networking-cisco in master are executed with newtron from mitaka-EOL. jobs with master and pike is not working. I think the only thing we can do is to pin it to the last built commit and remove the failed builds from delorean database.

Patch-set: 1
Label: Code-Review=0